Move all printing tests to PrintVertexTest
